divided realisation and declaration of sniffer
